+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AVRCP: new field `button_pressed` in `AVRCP_SUBEVENT_OPERATION`
 - AVRCP: `AVRCP_SUBEVENT_OPERATION` emitted for button release
 - AVRCP Controller: avrcp_controller_start_press_and_hold_cmd helps to support device buttons
+- AVRCP Controller: reassemble fragmented AVCTP packets
 - AVDTP: `avdtp_register_media_config_validator` allows to validate media codec configuration
 - A2DP Source: `ENABLE_A2DP_SOURCE_EXPLICIT_CONFIG` disables auto config. Requires call to `a2dp_source_set_config_{CODEC}'